Extract draft state into separate hook
