+++ client/packages/inventory/src/StockMovement/ListView/StockMovementModal.tsx
@@ -1,7 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
   Alert,
-  ArrayUtils,
   Box,
   Typography,
   useTranslation,
@@ -17,12 +16,8 @@ import { DialogButton } from '@common/components';
 import { useDialog } from '@common/hooks';
 import { FormControlLabel, Radio } from '@mui/material';
 import {
-  ItemStockOnHandFragment,
   LocationRowFragment,
   StockItemSearchInput,
-  StockLineRowFragment,
-  useLocationList,
-  useStockList,
 } from '@openmsupply-client/system';
 import {
   DraftStockMovementLine,
@@ -35,10 +30,11 @@ import {
   DraftStockMovementLineState,
   StockMovementLineTable,
 } from './StockMovementLineTable';
-
-type SelectionMode = 'byLocation' | 'byItem';
-
-type ItemOption = { id: string; code: string; name: string };
+import {
+  ItemOption,
+  SelectionMode,
+  useStockMovementDraft,
+} from './useStockMovementDraft';
 
 interface StockMovementModalProps {
   open: boolean;
@@ -46,56 +42,6 @@ interface StockMovementModalProps {
   mode: ModalMode | null;
   movement?: StockMovementRowFragment | null;
 }
-
-const lineFromStockLine = (
-  stockLine: StockLineRowFragment
-): DraftStockMovementLineState => {
-  const onHold = stockLine.onHold || (stockLine.location?.onHold ?? false);
-  return {
-    id: stockLine.id,
-    itemId: stockLine.itemId,
-    itemCode: stockLine.item.code,
-    itemName: stockLine.item.name,
-    restrictedLocationTypeId: stockLine.item.restrictedLocationTypeId,
-    fromStockLineId: stockLine.id,
-    fromLocationCode: stockLine.location?.code ?? null,
-    batch: stockLine.batch,
-    expiryDate: stockLine.expiryDate,
-    fromPackSize: stockLine.packSize,
-    availableNumberOfPacks: stockLine.availableNumberOfPacks,
-    onHold,
-    fromNumberOfPacks: undefined,
-    toLocation: null,
-    toPackSize: undefined,
-    toNumberOfPacks: undefined,
-  };
-};
-
-const lineFromMovement = (
-  movement: StockMovementRowFragment
-): DraftStockMovementLineState => {
-  const toPackSize = movement.toPackSize ?? movement.fromPackSize;
-  return {
-    id: movement.id,
-    itemId: '',
-    itemCode: movement.itemCode,
-    itemName: movement.itemName,
-    restrictedLocationTypeId: movement.restrictedLocationTypeId,
-    fromStockLineId: movement.fromStockLineId,
-    fromLocationCode: movement.fromLocation?.code ?? null,
-    batch: movement.batch,
-    expiryDate: movement.expiryDate,
-    fromPackSize: movement.fromPackSize,
-    availableNumberOfPacks: movement.availableNumberOfPacks,
-    onHold: movement.onHold || (movement.fromLocation?.onHold ?? false),
-    fromNumberOfPacks: movement.numberOfPacks,
-    toLocation: (movement.toLocation as LocationRowFragment | null) ?? null,
-    toPackSize,
-    toNumberOfPacks: toPackSize
-      ? (movement.numberOfPacks * movement.fromPackSize) / toPackSize
-      : undefined,
-  };
-};
 
 export const StockMovementModal = ({
   open,
@@ -110,17 +56,21 @@ export const StockMovementModal = ({
   const isEdit = mode === ModalMode.Update;
   const isDisabled = movement?.status === StockRelocationNodeStatus.Finalised;
 
-  const [selectionMode, setSelectionMode] =
-    useState<SelectionMode>('byLocation');
-  const [fromLocation, setFromLocation] = useState<LocationRowFragment | null>(
-    null
-  );
-  const [byItem, setByItem] = useState<ItemStockOnHandFragment | null>(null);
-  const [addedItemIds, setAddedItemIds] = useState<string[]>([]);
-  const [removedLineIds, setRemovedLineIds] = useState<string[]>([]);
-  const [edits, setEdits] = useState<
-    Record<string, Partial<DraftStockMovementLineState>>
-  >({});
+  const {
+    selectionMode,
+    switchMode,
+    fromLocation,
+    selectFromLocation,
+    byItem,
+    selectByItem,
+    locations,
+    itemOptions,
+    lines,
+    linesToMove,
+    onAddItem,
+    onUpdate,
+    onRemove,
+  } = useStockMovementDraft({ isEdit, movement });
 
   const { insert, isSaving } = useInsertStockMovement();
   const { update, isUpdating } = useUpdateStockMovement();
@@ -149,96 +99,6 @@ export const StockMovementModal = ({
     cancelButtonLabel: t('button.cancel'),
   });
 
-  const {
-    query: { data: locationData },
-  } = useLocationList(
-    {
-      sortBy: { key: 'name', direction: 'asc', isDesc: false },
-      first: 1000,
-    },
-    undefined,
-    !isEdit
-  );
-  const locations = locationData?.nodes ?? [];
-
-  const { data: locationStock } = useStockList(
-    {
-      filterBy: fromLocation
-        ? { location: { id: { equalTo: fromLocation.id } } }
-        : undefined,
-      first: 1000,
-    },
-    { enabled: !isEdit && selectionMode === 'byLocation' && !!fromLocation }
-  );
-
-  const { data: itemStock } = useStockList(
-    {
-      filterBy: byItem ? { itemId: { equalTo: byItem.id } } : undefined,
-      first: 1000,
-    },
-    { enabled: !isEdit && selectionMode === 'byItem' && !!byItem }
-  );
-
-  const clearSelection = () => {
-    setAddedItemIds([]);
-    setRemovedLineIds([]);
-    setEdits({});
-  };
-
-  const switchMode = (nextMode: SelectionMode) => {
-    setSelectionMode(nextMode);
-    setFromLocation(null);
-    setByItem(null);
-    clearSelection();
-  };
-
-  const fromStockNodes = (locationStock?.nodes ?? []).filter(
-    node =>
-      !!fromLocation &&
-      node.locationId === fromLocation.id &&
-      node.availableNumberOfPacks > 0
-  );
-  const itemStockNodes = (itemStock?.nodes ?? []).filter(
-    node =>
-      !!byItem && node.itemId === byItem.id && node.availableNumberOfPacks > 0
-  );
-
-  const sourceStockLines =
-    selectionMode === 'byLocation'
-      ? fromStockNodes.filter(node => addedItemIds.includes(node.itemId))
-      : itemStockNodes;
-
-  const lines: DraftStockMovementLineState[] = movement
-    ? [{ ...lineFromMovement(movement), ...edits[movement.id] }]
-    : sourceStockLines
-      .filter(stockLine => !removedLineIds.includes(stockLine.id))
-      .map(stockLine => ({
-        ...lineFromStockLine(stockLine),
-        ...edits[stockLine.id],
-      }));
-
-  const itemOptions = ArrayUtils.uniqBy(
-    fromStockNodes
-      .filter(node => !addedItemIds.includes(node.itemId))
-      .map(node => ({
-        id: node.itemId,
-        code: node.item.code,
-        name: node.item.name,
-      })),
-    'id'
-  ).sort((a, b) => a.name.localeCompare(b.name));
-
-  const onAddItem = (item: ItemOption) =>
-    setAddedItemIds(prev =>
-      prev.includes(item.id) ? prev : [...prev, item.id]
-    );
-
-  const onUpdate = (id: string, patch: Partial<DraftStockMovementLineState>) =>
-    setEdits(prev => ({ ...prev, [id]: { ...prev[id], ...patch } }));
-
-  const onRemove = (id: string) =>
-    setRemovedLineIds(prev => [...prev, id]);
-
   const resultingToPacks = (line: DraftStockMovementLineState) => {
     const toPackSize = line.toPackSize ?? line.fromPackSize;
     if (!toPackSize) return undefined;
@@ -255,10 +115,6 @@ export const StockMovementModal = ({
       NumUtils.isWholeNumber(toPacks)
     );
   };
-
-  const linesToMove = isEdit
-    ? lines
-    : lines.filter(line => !line.onHold && !!edits[line.id]);
 
   const hasFractionalPacks = linesToMove.some(line => {
     const toPacks = resultingToPacks(line);
@@ -429,16 +285,12 @@ export const StockMovementModal = ({
                 options={locations}
                 value={fromLocation}
                 getOptionLabel={location =>
-                  `${location.code} - ${location.name}${
-                    location.onHold ? ` (${t('label.on-hold')})` : ''
+                  `${location.code} - ${location.name}${location.onHold ? ` (${t('label.on-hold')})` : ''
                   }`
                 }
                 getOptionDisabled={location => location.onHold}
                 isOptionEqualToValue={(option, value) => option.id === value.id}
-                onChange={(_, location) => {
-                  setFromLocation(location);
-                  clearSelection();
-                }}
+                onChange={(_, location) => selectFromLocation(location)}
               />
             </Box>
             {fromLocation && (
@@ -469,10 +321,7 @@ export const StockMovementModal = ({
               currentItemId={byItem?.id}
               disabled={false}
               filter={{ hasStockOnHand: true }}
-              onChange={item => {
-                setByItem(item);
-                clearSelection();
-              }}
+              onChange={item => selectByItem(item)}
             />
           </Box>
         )}
